Merge develop into main

+++ modules/server.py
@@ -9,7 +9,7 @@ import shutil
 from uuid import uuid4
 import numpy as np
 import requests
-from nicegui import binding
+from nicegui import binding, ui
 
 from config import settings as mcssettings
 
@@ -312,36 +312,40 @@ class MinecraftServer:
 
     def delete(self, delete_dir: bool = False):
         """Deletes the server from servers.json"""
-        if delete_dir:
-            for item in os.listdir(self.server_path):
-                item_path = os.path.join(self.server_path, item)
-                try:
-                    if os.path.isfile(item_path) or os.path.islink(item_path):
-                        os.unlink(item_path)
-                    elif os.path.isdir(item_path):
-                        shutil.rmtree(item_path)
-                except Exception as e:
-                    print(f"Failed to delete {item_path}: {e}")
-            # remove server dir
-            shutil.rmtree(self.server_path)
+        if not self.running and not self.process:
+            if delete_dir:
+                for item in os.listdir(self.server_path):
+                    item_path = os.path.join(self.server_path, item)
+                    try:
+                        if os.path.isfile(item_path) or os.path.islink(item_path):
+                            os.unlink(item_path)
+                        elif os.path.isdir(item_path):
+                            shutil.rmtree(item_path)
+                    except Exception as e:
+                        print(f"Failed to delete {item_path}: {e}")
+                # remove server dir
+                shutil.rmtree(self.server_path)
 
-        # remove from server_list
-        assert self in server_list, "Invalid server"
-        server_list.remove(self)
+            # remove from server_list
+            assert self in server_list, "Invalid server"
+            server_list.remove(self)
 
-        # remove from global_settings
-        assert global_settings[self.uuid], "Invalid server"
-        del global_settings[self.uuid]
+            # remove from global_settings
+            assert global_settings[self.uuid], "Invalid server"
+            del global_settings[self.uuid]
 
-        # update settings
-        try:
-            self._save_settings()
+            # update settings
+            try:
+                self._save_settings()
 
-        except Exception as e:
-            print(f"Can't delete server: {e}")
-            raise
+            except Exception as e:
+                print(f"Can't delete server: {e}")
+                raise
 
-        print(f"Deleted server {self.uuid}")
+            print(f"Deleted server {self.uuid}")
+
+        else:
+            raise Exception("Can't delete the server while it's running.")
 
     def load_server_properties(self):
         """Loads server.properties file"""
@@ -356,16 +360,38 @@ class MinecraftServer:
                         key, value = line.strip().split("=", 1)
                         self.server_properties[key] = value
 
-    def save_server_properties(self):
-        """Saves server.properties file"""
-        raise NotImplementedError()
-        if self.has_server_properties and self.server_properties is not None:
+    def save_server_properties(self, editor: ui.editor) -> bool:
+        """
+        Saves server.properties file
+        Returns True if saved otherwise False
+        """
+        if self.has_server_properties and editor.content.get("json"):
+            self.server_properties = editor.content.get("json")
+
+            # Actually save server.properties file
+            # It's not a json, rows are NAME=VALUE
             with open(
                 os.path.join(self.server_path, "server.properties"),
                 mode="w",
                 encoding="utf-8",
             ) as properties:
-                self.server_properties = properties.readlines()
+                for setting, value in self.server_properties.items():
+                    properties.write(f"{setting}={value}\n")
+
+                properties.flush()
+                properties.close()
+
+            print(f"{self.uuid} properties saved")
+            return True
+
+        elif self.has_server_properties and editor.content.get("text"):
+            raise NotImplementedError("Editing in 'text' mode is currently disabled")
+            # The ugly
+            # editor.content.get("text").replace("\n", "").replace('"', "").replace(
+            #     "{", ""
+            # ).replace("}", "").split(",")[0].strip().split(": ")
+
+        return False
 
 
 def get_server_by_name(server_name: str) -> MinecraftServer | None:
